+++ tests/test_res1d_network.py
@@ -63,6 +63,7 @@ def test_data_item_dicts(test_file):
     assert len(res1d.catchments) == 0
     assert len(res1d.nodes) == 119
     assert len(res1d.reaches) == 118
+    assert len(res1d.structures) == 2
     assert len(res1d.global_data) == 0
 
 
@@ -185,3 +186,58 @@ def test_res1d_filter_readall(test_file_path):
     res1d = Res1D(test_file_path, nodes=nodes, reaches=reaches)
 
     res1d.read()
+
+
+def test_node_attributes(test_file):
+    res1d = test_file
+    nodes = res1d.nodes
+
+    nodes.n_1.WaterLevel.add()
+    nodes.n_2.WaterLevel.add()
+    df = res1d.read()
+
+    actual_max = round(df["WaterLevel:1"].max(), 3)
+    assert pytest.approx(actual_max) == 195.669
+
+    actual_max = round(df["WaterLevel:2"].max(), 3)
+    assert pytest.approx(actual_max) == 195.823
+
+
+def test_reach_attributes(test_file):
+    res1d = test_file
+    reaches = res1d.reaches
+
+    reaches.r_104l1.m_34_4131.WaterLevel.add()
+
+    reaches.r_9l1.m_10.WaterLevel.add()
+
+    reaches.r_100l1.m_23_8414.Discharge.add()
+
+    reaches.r_9l1.m_5.Discharge.add()
+
+    df = res1d.read()
+
+    actual_max = round(df["WaterLevel:104l1:34.4131"].max(), 3)
+    assert pytest.approx(actual_max) == 197.046
+
+    actual_max = round(df["WaterLevel:9l1:10"].max(), 3)
+    assert pytest.approx(actual_max) == 195.165
+
+    actual_max = round(df["Discharge:100l1:23.8414"].max(), 3)
+    assert pytest.approx(actual_max) == 0.1
+
+    actual_max = round(df["Discharge:9l1:5"].max(), 3)
+    assert pytest.approx(actual_max) == 0.761
+
+
+def test_structure_reach_attributes(test_file):
+    res1d = test_file
+    structures = res1d.structures
+
+    structures.s_119w1.Discharge.add()
+    structures.s_115p1.Discharge.add()
+
+    df = res1d.read()
+
+    max_discharge = round(df.max().max(), 3)
+    assert pytest.approx(max_discharge) == 1.491